Restaurant app

+++ app/src/main/java/com/kelvin/quickmenu/menu/itemsByCategory/model/RealtimeDatabase.kt
@@ -1,7 +1,5 @@
 package com.kelvin.quickmenu.menu.itemsByCategory.model
 
-import android.util.Log
-import android.widget.Toast
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.DatabaseException
@@ -29,9 +27,10 @@ class RealtimeDatabase {
                     if (snapshot.getValue() != null) {
                         for (n in snapshot.children) {
                             val item: ItemByCategory = n.getValue(ItemByCategory::class.java) as ItemByCategory
+                            //Verifying that the object is in the order list. If so, the same memory value is assigned
                             if (OrderSingleton.getItems().any { it.key.getName() == item.getName() } && item.getStatus()) {
                                 var sameObject=OrderSingleton.getItems().filter { it.key.getName()==item.getName() }.map{it.key}[0]
-                                sameObject.setPrice(item.getPrice().toDouble())
+                                sameObject.setPrice(item.getPrice())
                                 list.add(sameObject)
                             } else {
                                 item.setID(n.key!!.toInt())
